fix: Ensure numeric values are floats in upload error response
- Updated the upload_document function to convert confidence_score and min_threshold to float types before returning in the error response. This change ensures consistency in the data types returned for error handling related to document quality issues.

+++ backend/app/routers/upload.py
@@ -166,8 +166,8 @@ async def upload_document(
                         "error": "poor_document_quality",
                         "message": "Document quality is too low for reliable processing",
                         "details": {
-                            "confidence_score": round(confidence_score, 2),
-                            "min_threshold": min_confidence,
+                            "confidence_score": float(round(confidence_score, 2)),
+                            "min_threshold": float(min_confidence),
                             "issues": issues,
                             "suggestions": suggestions,
                         },
